+++ apps/auth/nextjs-ssr/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @crossmint/server-sdk-next-starter
 
+## 0.4.148
+
+### Patch Changes
+
+- Updated dependencies [4dc0dbf]
+- Updated dependencies [d31b60a]
+- Updated dependencies [27194e5]
+  - @crossmint/client-sdk-react-ui@2.6.17
+  - @crossmint/server-sdk@1.2.59
+
 ## 0.4.147
 
 ### Patch Changes