+++ packages/client/auth/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @crossmint/client-sdk-auth
 
+## 1.2.47
+
+### Patch Changes
+
+- Updated dependencies [27194e5]
+- Updated dependencies [4dc0dbf]
+  - @crossmint/common-sdk-base@0.9.17
+  - @crossmint/client-sdk-base@1.7.14
+  - @crossmint/common-sdk-auth@1.0.69
+
 ## 1.2.46
 
 ### Patch Changes